Changed all floats to doubles

+++ src/main/java/neat/ADD3I2O.java
@@ -16,52 +16,52 @@ public class ADD3I2O implements FitnessFunction {
 	}
 
 	@Override
-	public float calculateFitness(Genome gnm) {
-		List<Float> inList = new ArrayList<Float>();
-		inList.add(1f);
-		inList.add(1f);
-		inList.add(1f);
-		float d10 = 1f - gnm.calculate(inList).get(0);
-		float d11 = 1f - gnm.calculate(inList).get(1);
-		inList.set(0, 0f);
-		inList.set(1, 0f);
-		inList.set(2, 0f);
-		float d20 = gnm.calculate(inList).get(0);
-		float d21 = gnm.calculate(inList).get(1);
-		inList.set(0, 0f);
-		inList.set(1, 0f);
-		inList.set(2, 1f);
-		float d30 = 1f - gnm.calculate(inList).get(0);
-		float d31 = gnm.calculate(inList).get(1);
-		inList.set(0, 0f);
-		inList.set(1, 1f);
-		inList.set(2, 0f);
-		float d40 = 1f - gnm.calculate(inList).get(0);
-		float d41 = gnm.calculate(inList).get(1);
-		inList.set(0, 0f);
-		inList.set(1, 1f);
-		inList.set(2, 1f);
-		float d50 = gnm.calculate(inList).get(0);
-		float d51 = 1f - gnm.calculate(inList).get(1);
-		inList.set(0, 1f);
-		inList.set(1, 0f);
-		inList.set(2, 0f);
-		float d60 = 1f - gnm.calculate(inList).get(0);
-		float d61 = gnm.calculate(inList).get(1);
-		inList.set(0, 1f);
-		inList.set(1, 0f);
-		inList.set(2, 1f);
-		float d70 = gnm.calculate(inList).get(0);
-		float d71 = 1f - gnm.calculate(inList).get(1);
-		inList.set(0, 1f);
-		inList.set(1, 1f);
-		inList.set(2, 0f);
-		float d80 = gnm.calculate(inList).get(0);
-		float d81 = 1f - gnm.calculate(inList).get(1);
-		float d = d10 + d11 + d20 + d21 + d30 + d31 + d40 + d41 + d50 + d51 + d60 + d61 + d70 + d71 + d80 + d81;
-		return (16f - d) * (16f - d);
+	public double calculateFitness(Genome gnm) {
+		List<Double> inList = new ArrayList<Double>();
+		inList.add(1d);
+		inList.add(1d);
+		inList.add(1d);
+		double d10 = 1d - gnm.calculate(inList).get(0);
+		double d11 = 1d - gnm.calculate(inList).get(1);
+		inList.set(0, 0d);
+		inList.set(1, 0d);
+		inList.set(2, 0d);
+		double d20 = gnm.calculate(inList).get(0);
+		double d21 = gnm.calculate(inList).get(1);
+		inList.set(0, 0d);
+		inList.set(1, 0d);
+		inList.set(2, 1d);
+		double d30 = 1d - gnm.calculate(inList).get(0);
+		double d31 = gnm.calculate(inList).get(1);
+		inList.set(0, 0d);
+		inList.set(1, 1d);
+		inList.set(2, 0d);
+		double d40 = 1d - gnm.calculate(inList).get(0);
+		double d41 = gnm.calculate(inList).get(1);
+		inList.set(0, 0d);
+		inList.set(1, 1d);
+		inList.set(2, 1d);
+		double d50 = gnm.calculate(inList).get(0);
+		double d51 = 1d - gnm.calculate(inList).get(1);
+		inList.set(0, 1d);
+		inList.set(1, 0d);
+		inList.set(2, 0d);
+		double d60 = 1d - gnm.calculate(inList).get(0);
+		double d61 = gnm.calculate(inList).get(1);
+		inList.set(0, 1d);
+		inList.set(1, 0d);
+		inList.set(2, 1d);
+		double d70 = gnm.calculate(inList).get(0);
+		double d71 = 1d - gnm.calculate(inList).get(1);
+		inList.set(0, 1d);
+		inList.set(1, 1d);
+		inList.set(2, 0d);
+		double d80 = gnm.calculate(inList).get(0);
+		double d81 = 1d - gnm.calculate(inList).get(1);
+		double d = d10 + d11 + d20 + d21 + d30 + d31 + d40 + d41 + d50 + d51 + d60 + d61 + d70 + d71 + d80 + d81;
+		return (16d - d) * (16d - d);
 	}
-	public float getThreshold() {
-		return 225;
+	public double getThreshold() {
+		return 225d;
 	}
 }
